Extract "DataGrid export.customizeXlsxCell tests" module into a separate file
