+++ testing/tests/DevExpress.ui.widgets.dataGrid/dataGrid.export.customizeXlsxCell.tests.js
@@ -1,0 +1,44 @@
+import Helper from '../../helpers/DataGridExportHelper.js';
+
+Helper.QUnit_testStart();
+Helper.QUnit_module("DataGrid export.customizeXlsxCell tests");
+
+QUnit.test("customizeXlsxCell - set alignment: null for all xlsx cells", function(assert) {
+    const styles = Helper.STYLESHEET_HEADER_XML +
+        '<numFmts count="0"></numFmts>' +
+        Helper.BASE_STYLE_XML +
+        '<cellXfs count="7">' +
+        Helper.STYLESHEET_STANDARDSTYLES +
+        '<xf xfId="0" applyAlignment="1" fontId="0" applyNumberFormat="0" numFmtId="0"><alignment vertical="top" wrapText="0" horizontal="left" /></xf>' +
+        '<xf xfId="0" applyAlignment="1" fontId="1" applyNumberFormat="0" numFmtId="0"><alignment vertical="top" wrapText="0" horizontal="left" /></xf>' +
+        '<xf xfId="0" fontId="1" applyNumberFormat="0" numFmtId="0" />' +
+        '<xf xfId="0" fontId="0" applyNumberFormat="0" numFmtId="0" />' +
+        '</cellXfs>' +
+        Helper.STYLESHEET_FOOTER_XML;
+    const worksheet = Helper.WORKSHEET_HEADER_XML +
+        '<sheetPr/><dimension ref="A1:C1"/>' +
+        '<sheetViews><sheetView tabSelected="1" workbookViewId="0"><pane activePane="bottomLeft" state="frozen" ySplit="1" topLeftCell="A2" /></sheetView></sheetViews>' +
+        '<sheetFormatPr defaultRowHeight="15" outlineLevelRow="0" x14ac:dyDescent="0.25"/>' +
+        '<cols><col width="13.57" min="1" max="1" /></cols>' +
+        '<sheetData>' +
+        '<row r="1" spans="1:1" outlineLevel="0" x14ac:dyDescent="0.25"><c r="A1" s="5" t="s"><v>0</v></c></row>' +
+        '<row r="2" spans="1:1" outlineLevel="0" x14ac:dyDescent="0.25"><c r="A2" s="6" t="s"><v>1</v></c></row>' +
+        '</sheetData>' +
+        '<ignoredErrors><ignoredError sqref="A1:C2" numberStoredAsText="1" /></ignoredErrors></worksheet>';
+    const sharedStrings = Helper.SHARED_STRINGS_HEADER_XML + ' count="2" uniqueCount="2">' +
+        '<si><t>Field 1</t></si>' +
+        '<si><t>str1_1</t></si>' +
+        '</sst>';
+
+    Helper.runTest(
+        assert,
+        {
+            columns: [{ dataField: "field1" }],
+            dataSource: [{ field1: 'str1_1' }],
+            export: {
+                customizeXlsxCell: e => Object.assign(e.xlsxCell.style, { alignment: null })
+            }
+        },
+        { styles, worksheet, sharedStrings }
+    );
+});
